ready to deploy

+++ backend/server.js
@@ -27,25 +27,32 @@ const storage = multer.diskStorage({
 
 const upload = multer({ storage: storage })
 
-app.post('/upload', upload.single('image'), (req, res) => {
-    UserModel.create({ name: req.body.name, detail: req.body.detail, image: req.file.filename })
-        .then(result => res.json(result))
-        .catch(err => console.log(err))
+app.post('/upload', upload.single('image'), async (req, res) => {
+    const { name, detail } = req.body
+    const data = await UserModel.create({ name: name, detail: detail, image: req.file.filename })
+    if (data) res.status(201).json(data)
 })
-app.get('/', (req, res) => {
-    UserModel.find()
-        .then(result => res.status(200).send(result))
-        .catch(err => console.log(err))
+app.get('/', async (req, res) => {
+    const data = await UserModel.find()
+    if (data) res.status(200).json(data)
 })
-app.patch('/update/:id', upload.single('image'), async (req, res) => {
-    const update = req.body
+app.put('/update/:id', upload.single('image'), async (req, res) => {
+    const update = {}
+    if (req.body.name) update.name = req.body.name
+    if (req.body.detail) update.name = req.body.detail
+    if (req.file) update.image = req.file.filename
+    if (Object.keys(update).length === 0 && !req.file) {
+        return res.status(400).send({ message: "fill atleast one field for update" })
+    }
     const old = await UserModel.findById({ _id: req.params.id })
     const upd = await UserModel.findOneAndUpdate({ _id: req.params.id }, { $set: update }, { new: true })
     if (!upd) res.status(404).send("error in update")
-    res.send(upd)
+    res.status(200).send(upd)
     try {
-        const imgpath = `./public/images/${old.image}`
-        await fs.promises.unlink(imgpath)
+        if (update.image) {
+            const imgpath = `./public/images/${old.image}`
+            await fs.promises.unlink(imgpath)
+        }
     } catch (error) {
 
     }
